+++ CHANGELOG.md
@@ -1,5 +1,65 @@
 # @contentauth/c2pa-node
 
+## 0.5.5
+
+### Patch Changes
+
+- 1bdc7bf: Update trustmark model url
+- f40b882: Fix x86 mac release. Allow custom index.node
+- 4ed8b65: Add method for adding redactions to builder
+
+## 0.5.4
+
+### Patch Changes
+
+- c5e18ed: Allow partial verify settings
+
+## 0.5.3
+
+### Patch Changes
+
+- c94044e: export Ingredient from c2pa-types
+
+## 0.5.2
+
+### Patch Changes
+
+- ec86837: Update c2pa crate to 0.77.0
+
+## 0.5.1
+
+### Patch Changes
+
+- df6a49c: Add context to additional methods. Update c2pa crate.
+
+## 0.5.0
+
+### Minor Changes
+
+- 70a7ee3: Add settings to Builder and Reader constructors, remove global settings
+
+## 0.4.2
+
+### Patch Changes
+
+- 26f0c35: Fix settings on tasks. Fix Builder toArchive
+
+## 0.4.1
+
+### Patch Changes
+
+- c6defcc: Add addIngredientFromReader method
+
+## 0.4.0
+
+### Minor Changes
+
+- 9e2fb33: Integrate postCawgValidate into Reader.
+
+### Patch Changes
+
+- 8a87c6d: Make asset in Builder.addIngredient optional. Add Builder.addAction
+
 ## 0.3.0
 
 ### Minor Changes